API adjustments for long -> int transition

+++ src/main/java/com/jug/segmentation/ParaMaxFlow.java
@@ -132,14 +132,14 @@ public class ParaMaxFlow< T extends RealType< T > > {
 		final float eps = 0.0000001f;
 
 		// for each pixel in input image --> create unary term
-		for ( long z = 0; z < dims[ 2 ]; z++ ) {
-			for ( long y = 0; y < dims[ 1 ]; y++ ) {
-				for ( long x = 0; x < dims[ 0 ]; x++ ) {
+		for ( int z = 0; z < dims[ 2 ]; z++ ) {
+			for ( int y = 0; y < dims[ 1 ]; y++ ) {
+				for ( int x = 0; x < dims[ 0 ]; x++ ) {
 					raUnaryPotentialImage.setPosition( new long[] { x, y, z } );
 
 					final double likelihood = unaryCostFactor * raUnaryPotentialImage.get().getRealDouble();
 
-					pixelId = ( int ) ( z * dims[ 1 ] * dims[ 0 ] + y * dims[ 0 ] + x );
+					pixelId = (int)(z * dims[ 1 ] * dims[ 0 ] + y * dims[ 0 ] + x);
 					parametric.addUnaryTerm( pixelId, 1.0, likelihood );
 				}
 			}
@@ -164,7 +164,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 
 						cost += isingCost;
 
-						final long xNeighborId = pixelId + 1;
+						final int xNeighborId = pixelId + 1;
 						parametric.addPairwiseTerm( pixelId, xNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 					if ( y + 1 < dims[ 1 ] ) {
@@ -175,7 +175,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 
 						cost += isingCost;
 
-						final long yNeighborId = pixelId + dims[ 0 ];
+						final int yNeighborId = pixelId + (int)(dims[ 0 ]);
 						parametric.addPairwiseTerm( pixelId, yNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 					// connect in z-direction ONLY if is3d==TRUE!!!
@@ -187,7 +187,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 
 						cost += isingCost;
 
-						final long zNeighborId = pixelId + dims[ 0 ] * dims[ 1 ];
+						final int zNeighborId = pixelId + (int)(dims[ 0 ] * dims[ 1 ]);
 						parametric.addPairwiseTerm( pixelId, zNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 				}
@@ -278,7 +278,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 							cost *= ( 1.0 - raProbMap.get().getRealDouble() );
 						}
 
-						final long xNeighborId = pixelId + 1;
+						final int xNeighborId = pixelId + 1;
 						parametric.addPairwiseTerm( pixelId, xNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 					if ( y + 1 < dims[ 1 ] ) {
@@ -294,7 +294,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 							cost *= ( 1.0 - raProbMap.get().getRealDouble() );
 						}
 
-						final long yNeighborId = pixelId + dims[ 0 ];
+						final int yNeighborId = pixelId + (int)(dims[ 0 ]);
 						parametric.addPairwiseTerm( pixelId, yNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 					// connect in z-direction ONLY if is3d==TRUE!!!
@@ -311,7 +311,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 							cost *= ( 1.0 - raProbMap.get().getRealDouble() );
 						}
 
-						final long zNeighborId = pixelId + dims[ 0 ] * dims[ 1 ];
+						final int zNeighborId = pixelId + (int)(dims[ 0 ] * dims[ 1 ]);
 						parametric.addPairwiseTerm( pixelId, zNeighborId, 0.0, cost, cost, 0.0 ); // add term with costs E00, E01, E10, and E11,
 					}
 				}
@@ -342,7 +342,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 		dims = new long[ ivRet.numDimensions() ];
 		ivRet.dimensions( dims );
 
-		for ( long graphNodeId = 0; graphNodeId < Views.iterable( rai ).size(); graphNodeId++ ) {
+		for ( int graphNodeId = 0; graphNodeId < Views.iterable( rai ).size(); graphNodeId++ ) {
 			final long numRegions = parametric.getRegionCount( parametric.getRegion( graphNodeId ) );
 
 			final long z = graphNodeId / ( dims[ 0 ] * dims[ 1 ] );
@@ -374,7 +374,7 @@ public class ParaMaxFlow< T extends RealType< T > > {
 		}
 		final RandomAccess< BitType > raRet = ivRet.randomAccess();
 
-		for ( long graphNodeId = 0; graphNodeId < Views.iterable( rai ).size(); graphNodeId++ ) {
+		for ( int graphNodeId = 0; graphNodeId < Views.iterable( rai ).size(); graphNodeId++ ) {
 			final long numRegions = parametric.getRegionCount( parametric.getRegion( graphNodeId ) );
 
 			final long z = graphNodeId / ( dims[ 0 ] * dims[ 1 ] );
